requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.1
cssselect==1.0.3
Flask==1.0.1
gunicorn==19.8.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.0
MarkupSafe==1.0
marshmallow==2.15.1
marshmallow-jsonapi==0.17.0
parsel==1.4.0
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
python-decouple==3.1
python-scrapyd-api==2.0.1
pytz==2018.3
queuelib==1.5.0
requests==2.18.4
Scrapy==1.5.0
scrapyd==1.2.0
selenium==3.11.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
Werkzeug==0.14.1
zope.interface==4.4.3
